Move RemoteDataStore interface into persistence.remote
