+++ gnd/src/main/java/com/google/android/gnd/GndApplicationModule.java
@@ -21,7 +21,7 @@ import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.android.gnd.inject.ActivityScoped;
 import com.google.android.gnd.persistence.local.LocalDataStore;
 import com.google.android.gnd.persistence.local.room.RoomDataStore;
-import com.google.android.gnd.service.RemoteDataStore;
+import com.google.android.gnd.persistence.remote.RemoteDataStore;
 import com.google.android.gnd.service.firestore.FirestoreDataStore;
 import com.google.android.gnd.ui.common.ViewModelModule;
 import dagger.Binds;
