+++ gnd/src/main/java/com/google/android/gnd/service/firestore/FirestoreDataStore.java
@@ -19,7 +19,7 @@ package com.google.android.gnd.service.firestore;
 import androidx.annotation.Nullable;
 import com.google.android.gnd.persistence.DataStoreEvent;
 import com.google.android.gnd.rx.RxTask;
-import com.google.android.gnd.service.RemoteDataStore;
+import com.google.android.gnd.persistence.remote.RemoteDataStore;
 import com.google.android.gnd.system.AuthenticationManager.User;
 import com.google.android.gnd.vo.Feature;
 import com.google.android.gnd.vo.FeatureUpdate.RecordUpdate.ResponseUpdate;
